requirements.txt
MetaTrader5==5.0.5640
pandas>=2.2.0
pandas-ta==0.4.71b0
numpy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
anthropic>=0.21.0
fastapi>=0.110.0
uvicorn>=0.27.1
pytest>=8.0.0
